+++ source/common/json/json_streamer.h
@@ -449,5 +449,10 @@ private:
  */
 using Streamer = StreamerBase<BufferOutput>;
 
+/**
+ * A Streamer that streams to a string.
+ */
+using StringStreamer = StreamerBase<StringOutput>;
+
 } // namespace Json
 } // namespace Envoy
