+++ source/common/protobuf/utility.h
@@ -685,13 +685,6 @@ public:
    * @return wrapped list value.
    */
   static ProtobufWkt::Value listValue(const std::vector<ProtobufWkt::Value>& values);
-
-  /**
-   * Convert a ProtobufWkt::Value to a JSON string.
-   * @param value message of type type.googleapis.com/google.protobuf.Value
-   * @param dest JSON string.
-   */
-  static void appendJsonToString(const ProtobufWkt::Value& value, std::string& dest);
 };
 
 /**
