+++ test/common/protobuf/utility_test.cc
@@ -1358,102 +1358,6 @@ TEST_F(ProtobufUtilityTest, ValueUtilLoadFromYamlObjectWithIgnoredEntries) {
       "struct_value { fields { key: \"baz\" value { string_value: \"qux\" } } }"));
 }
 
-std::string toJson(const ProtobufWkt::Value& v) {
-  std::string json_string;
-  ValueUtil::appendJsonToString(v, json_string);
-  return json_string;
-}
-
-TEST_F(ProtobufUtilityTest, ValueToJson) {
-  {
-    // null
-    ProtobufWkt::Value v;
-    EXPECT_EQ(toJson(v), "null");
-
-    v.set_null_value(ProtobufWkt::NULL_VALUE);
-    EXPECT_EQ(toJson(v), "null");
-  }
-
-  {
-    // bool
-    ProtobufWkt::Value v;
-    v.set_bool_value(true);
-    EXPECT_EQ(toJson(v), "true");
-
-    v.set_bool_value(false);
-    EXPECT_EQ(toJson(v), "false");
-  }
-
-  {
-    // number
-    ProtobufWkt::Value v;
-
-    v.set_number_value(1);
-    EXPECT_EQ(toJson(v), "1");
-
-    v.set_number_value(1.1);
-    EXPECT_EQ(toJson(v), "1.1");
-  }
-
-  {
-    // string
-    ProtobufWkt::Value v;
-
-    v.set_string_value("foo");
-    EXPECT_EQ(toJson(v), "\"foo\"");
-  }
-
-  {
-    // struct
-    ProtobufWkt::Value v;
-    auto* struct_value = v.mutable_struct_value();
-    EXPECT_EQ(toJson(v), R"EOF({})EOF");
-
-    struct_value->mutable_fields()->insert({"foo", ValueUtil::stringValue("bar")});
-    EXPECT_EQ(toJson(v), R"EOF({"foo":"bar"})EOF");
-  }
-
-  {
-    // list
-    ProtobufWkt::Value v;
-    auto* list_value = v.mutable_list_value();
-
-    EXPECT_EQ(toJson(v), R"EOF([])EOF");
-
-    list_value->add_values()->set_string_value("foo");
-    list_value->add_values()->set_string_value("bar");
-
-    EXPECT_EQ(toJson(v), R"EOF(["foo","bar"])EOF");
-  }
-
-  {
-    // Complex structure
-    const std::string yaml = R"EOF(
-    a:
-      a:
-        - a: 1
-          b: 2
-      b:
-        - a: 3
-          b: 4
-        - a: 5
-      c: true
-      d: [5, 3.14]
-      e: foo
-      f: 1.1
-    b: [1, 2, 3]
-    c: bar
-    )EOF";
-
-    ProtobufWkt::Value v;
-    MessageUtil::loadFromYaml(yaml, v, ProtobufMessage::getNullValidationVisitor());
-
-    EXPECT_EQ(
-        toJson(v),
-        R"EOF({"a":{"a":[{"a":1,"b":2}],"b":[{"a":3,"b":4},{"a":5}],"c":true,"d":[5,"3.14"],"e":"foo","f":"1.1"},"b":[1,2,3],"c":"bar"})EOF");
-  }
-}
-
 TEST(LoadFromYamlExceptionTest, BadConversion) {
   std::string bad_yaml = R"EOF(
 admin:
